+++ m68000/src/lib.rs
@@ -86,10 +86,13 @@
 //!
 //! See the `CINTERFACE.md` file in the repo for more information.
 //!
-//! ## TODO:
+//! ## Potential caveats
+//! - ABCD/NBCD/SBCD/DIVS/DIVU may not always procuce the correct output.
+//! - DIVS/DIVU always execute using their maximum execution time.
+//! - Long exception stack frame is not implemented.
+//!
+//! ## TODO
 //! - Let memory access return extra read or write cycles for accuracy.
-//! - Verify ABCD, NBCD, SBCD, DIVS and DIVU instructions.
-//! - Implement long exception stack frame.
 
 #![feature(bigint_helper_methods)]
 #![feature(btree_drain_filter)]